Energy Consumption in the Steel Industry
Final paper for DSP 557. 

This paper is on the predictability of energy consumption (in kWh) of the steel industry. It analyzes publicly available data from a Korean steel firm. The primary aim of this project were to determine if regression models could be used to predict energy consumption. The secondary aims were to see if the steel industry's energy usage was cyclical in nature, if energy usage could be predicted by date and time related features alone, if adding additional features defining time and date improved model accuracy, and if any of the specific added time and date related features had a significant impact on model accuracy.